blog/indexをcomponentに切り出す
